# docs/guides/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Project Manager UI API
  description: |
    REST API for the Project Manager UI — a real-time project management dashboard
    for stakeholders and AI agents. Built with Next.js 15, PostgreSQL 16, Prisma, and SSE.

    **Base URL**: `http://localhost:3000/api` (local) or your deployment host.

    **Authentication** (dual):
    - **Agent access**: `X-API-Key` header (see [ApiKeyAuth](#/components/securitySchemes/ApiKeyAuth))
    - **Browser user access**: `auth_token` cookie set by login/register (see [CookieAuth](#/components/securitySchemes/CookieAuth))
    
    Write operations (POST, PUT, DELETE) require authentication. Read operations (GET) 
    are publicly accessible for backward compatibility (planned enforcement in v2).

    The SSE endpoint (`/api/events`) requires authentication as of v1.1.0.
  version: "1.1.0"
  contact:
    name: Project Manager UI Team
    url: https://github.com/gansru/ui_pm
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://your-domain.com/api
    description: Production server
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for AI agent authentication. Valid for all authenticated operations.
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: JWT cookie for browser user authentication. Set automatically by login/register endpoints. httpOnly, sameSite=lax, 7-day expiry.

    # ─── Request Payloads ───
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
        - confirmPassword
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Display name (required)
          example: "Roman Govorov"
        email:
          type: string
          format: email
          maxLength: 255
          description: Email address (must not already exist)
          example: "roman@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 128
          description: Password (min 8 characters, cannot be a common password)
          example: "MyStr0ng!P@ss"
        confirmPassword:
          type: string
...
  schemas:
...
    ErrorBody:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorResponse"
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          examples:
            - VALIDATION_ERROR
            - NOT_FOUND
            - CONFLICT
            - UNAUTHORIZED
            - RATE_LIMITED
            - INTERNAL_ERROR
        message:
          type: string
          description: Human-readable error description
        details:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
                description: Dot-path of the invalid field
              message:
                type: string
                description: Validation error message
          description: Detailed validation errors (present only for VALIDATION_ERROR)
...
    PaginatedResponseTask:
      type: object
      required:
        - data
        - total
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        total:
          type: integer
          description: Total number of records returned
    PaginatedResponseUser:
      type: object
      required:
        - data
        - total
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
          description: Total number of records returned

    # ---- Entity Schemas ----
...
    Task:
      type: object
      required:
        - id
        - projectId
        - title
        - status
        - priority
        - assignee
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique task identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        projectId:
          type: string
          format: uuid
          description: Associated project ID
        subprojectId:
          type: string
          format: uuid
          nullable: true
          description: Associated subproject ID (optional)
        title:
          type: string
          maxLength: 200
          description: Task title
          example: "Implement login page"
        description:
          type: string
          nullable: true
          maxLength: 2000
          description: Detailed task description
        status:
          type: string
          enum:
            - in_work
            - review
            - done
          description: Current workflow status
          example: "in_work"
        priority:
          type: string
          enum:
            - high
            - medium
            - low
          description: Task priority level
          example: "medium"
        assignee:
          type: string
          maxLength: 100
          description: Name of assigned user
          example: "Roman Govorov"
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 last update timestamp
    User:
      type: object
      required:
        - id
        - name
        - role
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
        name:
          type: string
          maxLength: 100
          description: Display name
          example: "Roman Govorov"
        role:
          type: string
          enum:
            - stakeholder
            - agent
          description: User role
          example: "stakeholder"
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 creation timestamp

    # ---- Request Payloads ----
...
    CreateTaskRequest:
      type: object
      required:
        - projectId
        - title
        - assignee
      properties:
        projectId:
          type: string
          format: uuid
          description: UUID of the parent project
          example: "550e8400-e29b-41d4-a716-446655440000"
        subprojectId:
          type: string
          format: uuid
          nullable: true
          description: Optional subproject assignment
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title (required)
          example: "Implement authentication flow"
        description:
          type: string
          maxLength: 2000
          description: Optional task description
        priority:
          type: string
          enum:
            - high
            - medium
            - low
          default: medium
          description: Task priority (defaults to medium)
          example: "high"
        assignee:
          type: string
          minLength: 1
          maxLength: 100
          description: Assignee display name (required)
          example: "Agent-T67"
        status:
          type: string
          enum:
            - in_work
            - review
            - done
          default: in_work
          description: Task initial status (defaults to in_work)
          example: "in_work"
...
    CreateUserRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: User display name (required)
          example: "Jane Smith"
        role:
          type: string
          enum:
            - stakeholder
            - agent
          default: stakeholder
          description: User role (defaults to stakeholder)
        apiKey:
          type: string
          minLength: 32
          maxLength: 255
          description: Optional API key for agent authentication
...
paths:
...
  /tasks:
    get:
      tags:
        - tasks
      summary: List tasks
      description: |
        Returns a filtered list of tasks. Supports query parameters for filtering.
        Tasks are ordered by priority (high first), then creation date (newest first).
        No authentication required.
      operationId: listTasks
      parameters:
        - name: projectId
          in: query
          description: Filter by project UUID
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: Filter by task status
          required: false
          schema:
            type: string
            enum:
              - in_work
              - review
              - done
      responses:
        "200":
          description: A list of tasks matching filters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponseTask"
              example:
                data:
                  - id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    projectId: "550e8400-e29b-41d4-a716-446655440000"
                    subprojectId: null
                    title: "Implement login flow"
                    description: "OAuth2 + JWT implementation"
                    status: in_work
                    priority: high
                    assignee: "Agent-T67"
                    createdAt: "2026-08-01T10:00:00.000Z"
                    updatedAt: "2026-08-10T12:00:00.000Z"
                total: 1
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    post:
      tags:
        - tasks
      summary: Create a new task
      description: |
        Creates a new task within a project. Priority defaults to `medium`. Status defaults to `in_work` but can be set explicitly via the `status` field (one of: `in_work`, `review`, `done`).
        Requires API key authentication. Emits a `task_created` SSE event.
      operationId: createTask
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
            example:
              projectId: "550e8400-e29b-41d4-a716-446655440000"
              title: "Implement authentication flow"
              description: "JWT tokens + refresh tokens"
              priority: high
              status: review
              assignee: "Agent-T67"
      responses:
        "201":
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Validation error or invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          description: Foreign key constraint (invalid project ID)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
...
  /users:
    get:
      tags:
        - users
      summary: List users
      description: |
        Returns a list of all users, ordered alphabetically by name.
        API keys are intentionally excluded from the response for security.
        No authentication required.
      operationId: listUsers
      responses:
        "200":
          description: A list of users (without API keys)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponseUser"
              example:
                data:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    name: "Roman Govorov"
                    role: stakeholder
                    createdAt: "2026-08-01T10:00:00.000Z"
                  - id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    name: "Agent-T67"
                    role: agent
                    createdAt: "2026-08-01T10:00:00.000Z"
                total: 2
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    post:
      tags:
        - users
      summary: Create a user
      description: |
        Creates a new user account. An optional API key can be provided for agent accounts.
        Requires API key authentication.
      operationId: createUser
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
            example:
              name: "Jane Smith"
              role: stakeholder
      responses:
        "201":
          description: User created successfully (excludes API key from response)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error or invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          description: Duplicate user (unique constraint on name)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"

  # ============================================================
  # EVENTS (SSE) — /api/events
  # ============================================================
...